# Conflicts:
#	build.gradle
#	buildconfig/dependencies/dependencies.gradle
#	buildconfig/kotlin/core-module.gradle
#	buildconfig/kotlin/module.gradle
#	customerapp/build.gradle
#	dynamic-features/content/df_createpost/build.gradle
#	dynamic-features/content/df_imagepicker_insta/build.gradle
#	features/communication/attachvoucher/build.gradle
#	features/content/play/build.gradle
#	features/content/play_broadcaster/build.gradle
#	features/content/play_common/build.gradle
#	features/discovery/autocomplete/build.gradle
#	features/discovery/autocomplete_component/build.gradle
#	features/discovery/home/build.gradle
#	features/discovery/home_component/build.gradle
#	features/discovery/home_nav/build.gradle
#	features/discovery/recentview/build.gradle
#	features/discovery/search/build.gradle
#	features/fintech/pdp_fintech_widget/build.gradle
#	features/home/searchbar/build.gradle
#	features/media/live-broadcaster/build.gradle
#	features/media/mediapicker/build.gradle
#	features/merchant/brandlist/build.gradle
#	features/merchant/official_store/build.gradle
#	features/ordermanagement/snapshot/build.gradle
#	features/payment/gopay.kyc/build.gradle
#	features/payment/pms/build.gradle
#	features/recharge/digital_checkout/build.gradle
#	features/recharge/recharge_component/build.gradle
#	features/transaction/buyerorder/build.gradle
#	features/transaction/checkout/build.gradle
#	features/transaction/unifyorderhistory/build.gradle
#	features/transaction/wishlist/build.gradle
#	features/user/settingnotif/build.gradle
#	libraries/akamai_bot_lib/build.gradle
#	libraries/analytics_debugger/build.gradle
#	libraries/applink/build.gradle
#	libraries/composite_lib/analytic_param/build.gradle
#	libraries/composite_lib/analytic_testapp/build.gradle
#	libraries/composite_lib/annotation_processor_analytic/build.gradle
#	libraries/composite_lib/buildSrc/build.gradle
#	libraries/composite_lib/gqlquery_annotation/build.gradle
#	libraries/composite_lib/gqlquery_annotation_processor/build.gradle
#	libraries/device_info/build.gradle
#	libraries/inappupdate/build.gradle
#	libraries/logger/build.gradle
#	libraries/macrobenchmark/build.gradle
#	libraries/tokopatch/build.gradle
#	libraries/track/build.gradle
#	libraries/utils/build.gradle
#	settings.gradle
#	testing/joint/dg_transaction/build.gradle
#	testing/joint/inactivephone_otp/build.gradle
#	testing/joint/pdp_checkout/build.gradle

+++ features/user/chatbot/src/main/java/com/tokopedia/chatbot/domain/pojo/csatRating/websocketCsatRatingResponse/WebSocketCsatResponse.kt
@@ -49,7 +49,7 @@ data class WebSocketCsatResponse(
 	val toBuyer: Boolean? = null,
 
 	@SerializedName("msg_id")
-	val msgId: Int? = null,
+	val msgId: Long? = null,
 
 	@SerializedName("is_bot")
 	val isBot: Boolean? = null
@@ -69,7 +69,7 @@ data class WebSocketCsatResponse(
 			parcel.readValue(Int::class.java.classLoader) as? Int,
 			parcel.readString(),
 			parcel.readValue(Boolean::class.java.classLoader) as? Boolean,
-			parcel.readValue(Int::class.java.classLoader) as? Int,
+			parcel.readValue(Long::class.java.classLoader) as? Long,
 			parcel.readValue(Boolean::class.java.classLoader) as? Boolean) {
 	}
